Input: normalize project
