+++ make/CMakeLists_bgfx-linux-aarch64.txt
@@ -61,8 +61,6 @@ add_executable(vpinball
    src/parts/dragpoint.cpp
    src/utils/eventproxy.h
    src/parts/backGlass.cpp
-   src/core/pininput.cpp
-   src/core/pininput_OpenPinDev.cpp
    src/parts/pinbinary.cpp
    src/core/vpversion.h
 
@@ -80,7 +78,6 @@ add_executable(vpinball
    src/core/ieditable.h
    src/core/iselect.cpp
    src/core/iselect.h
-   src/core/pininput.h
    src/core/pinundo.cpp
    src/core/pinundo.h
    src/core/main.cpp
@@ -100,7 +97,10 @@ add_executable(vpinball
    src/core/VPXPluginAPIImpl.h
    src/core/VPXPluginAPIImpl.cpp
 
+   src/input/pininput.h
+   src/input/pininput.cpp
    src/input/OpenPinDevHandler.h
+   src/input/OpenPinDevHandler.cpp
    src/input/SDLInputHandler.h
 
    src/math/bbox.h
