+++ make/CMakeLists_bgfx-macos-x64.txt
@@ -52,7 +52,6 @@ add_executable(vpinball
    src/parts/dragpoint.cpp
    src/utils/eventproxy.h
    src/parts/backGlass.cpp
-   src/core/pininput.cpp
    src/parts/pinbinary.cpp
    src/core/vpversion.h
 
@@ -70,8 +69,6 @@ add_executable(vpinball
    src/core/ieditable.h
    src/core/iselect.cpp
    src/core/iselect.h
-   src/core/pininput.h
-   src/core/pininput_OpenPinDev.cpp
    src/core/pinundo.cpp
    src/core/pinundo.h
    src/core/main.cpp
@@ -91,7 +88,10 @@ add_executable(vpinball
    src/core/VPXPluginAPIImpl.h
    src/core/VPXPluginAPIImpl.cpp
 
+   src/input/pininput.h
+   src/input/pininput.cpp
    src/input/OpenPinDevHandler.h
+   src/input/OpenPinDevHandler.cpp
    src/input/SDLInputHandler.h
 
    src/math/bbox.h
